refactor: SMALLINT 컬럼 정합 위해 Integer → Short 변환

+++ backend/src/main/java/com/ssafy/s309/domain/user/dto/GuardianResponse.java
@@ -3,7 +3,7 @@ package com.ssafy.s309.domain.user.dto;
 import com.ssafy.s309.domain.user.entity.WardGuardian;
 
 public record GuardianResponse(
-    Integer id, Integer wardId, Integer guardianId, String relation, Integer priority) {
+    Integer id, Integer wardId, Integer guardianId, String relation, Short priority) {
 
   public static GuardianResponse from(WardGuardian wg) {
     return new GuardianResponse(
